pvh/arch: Introduce EntryPoint struct
In order to properly configure the initial vCPU register state
and boot parameters in guest memory, we must specify which
boot protocol to use with the kernel entry point address.
On x86-64 (the only architecture where multiple boot protocols
are supported) we print the protocol used to load the kernel
at the debug log level.

Create an EntryPoint struct that contains the required
information. This structure will later be used in the vCPU
configuration methods to set the appropriate initial
conditions for the guest.

This commit also splits the load_kernel function into an x86-64
specific version and an aarch64 specific version.

Signed-off-by: Colin Percival <cperciva@freebsd.org>
Co-authored-by: Alejandro Jimenez <alejandro.j.jimenez@oracle.com>